requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
idna==3.4
python-decouple==3.8
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
